refactor(summarylist): rewrite to agreed format
