+++ source/components/organisms/SummaryList/SummaryList.tsx
@@ -1,10 +1,6 @@
 import React, { useMemo } from "react";
-import PropTypes from "prop-types";
-import styled from "styled-components/native";
 
 import { GroupedList } from "../../molecules";
-
-import { Text, Heading } from "../../atoms";
 
 import SummaryListItemComponent from "./SummaryListItem";
 
@@ -12,123 +8,20 @@ import { isObject, deepCopy } from "../../../helpers/Objects";
 
 import { getValidColorSchema } from "../../../theme/themeHelpers";
 
-import type { PrimaryColor, ThemeType } from "../../../theme/themeHelpers";
-import type { InputType } from "../../atoms/Input/Input";
-import type { Help } from "../../../types/FormTypes";
+import {
+  SumLabel,
+  SumText,
+  SumContainer,
+  SummarySpacer,
+} from "./SummaryList.styled";
 
-const SumLabel = styled(Heading)<{ colorSchema: string }>`
-  margin-top: 5px;
-  margin-left: 3px;
-  font-weight: ${(props) => props.theme.fontWeights[1]};
-  font-size: ${(props) => props.theme.fontSizes[3]}px;
-  color: ${(props) => props.theme.colors.primary[props.colorSchema][1]};
-`;
-const SumText = styled(Text)`
-  margin-left: 4px;
-  margin-top: 10px;
-`;
-// TODO: the sum component should be sent as a footer to the grouped list, at which point this container should be removed.
-// Currently it's using a negative margin, which is a hack and only meant as a temporary fix.
-const SumContainer = styled.View<{ colorSchema: string }>`
-  width: 100%;
-  height: auto;
-  border-radius: 9.5px;
-  overflow: hidden;
-  margin-bottom: 24px;
-  margin-top: -64px;
-  padding-bottom: 20px;
-  padding-top: 16px;
-  padding-left: 16px;
-  padding-right: 16px;
-  background: ${(props) =>
-    props.theme.colors.complementary[props.colorSchema][3]};
-`;
-
-interface SummarySpacerProps {
-  colorSchema: PrimaryColor;
-  theme: ThemeType;
-}
-
-const SummarySpacer = styled.View<SummarySpacerProps>`
-  margin-bottom: 16px;
-  height: 4px;
-  width: 32px;
-
-  background: ${(props) =>
-    props.theme.label.colors[props.colorSchema].underline};
-`;
-
-export interface SummaryListItem {
-  title: string;
-  id: string;
-  type:
-    | "number"
-    | "text"
-    | "date"
-    | "checkbox"
-    | "arrayNumber"
-    | "arrayText"
-    | "arrayDate"
-    | "editableListText"
-    | "editableListNumber"
-    | "editableListDate"
-    | "spacer";
-  category?: string;
-  inputId?: string;
-  inputSelectValue?: InputType;
-  fieldStyle?: string;
-}
-
-interface SummaryListCategory {
-  category: string;
-  description: string;
-  sortField?: string;
-}
-
-interface Props {
-  heading: string;
-  items: SummaryListItem[];
-  categories?: SummaryListCategory[];
-  onChange: (
-    answers: Record<string, any> | string | number | boolean,
-    fieldId: string
-  ) => void;
-  onBlur: (
-    answers: Record<string, any> | string | number | boolean,
-    fieldId: string
-  ) => void;
-  colorSchema: PrimaryColor;
-  answers: Record<string, any>;
-  validationErrors?: Record<
-    string,
-    | { isValid: boolean; message: string }
-    | Record<string, { isValid: boolean; message: string }>[]
-  >;
-  showSum: boolean;
-  startEditable?: boolean;
-  help?: Help;
-  editable?: boolean;
-}
-
-type Answer = {
-  otherassetDescription?: string;
-  text?: string;
-  description?: string;
-} & Record<string, string | boolean | number>;
-
-interface List {
-  items: SummaryListItem[];
-  answers: Answer[] | string | number | boolean;
-}
-
-type Item = {
-  item: SummaryListItem;
-  value?: string | boolean | number;
-  index: number;
-  otherassetDescription?: string;
-  text?: string;
-  description?: string;
-} & Record<string, unknown>;
+import type {
+  SummaryListItem,
+  Props,
+  Answer,
+  List,
+  Item,
+} from "./SummaryList.types";
 
 const ArrayType = ["arrayNumber", "arrayText", "arrayDate"];
 
@@ -161,20 +54,22 @@ function findLastIndex<T>(
  * The things to summarize is specified in the items prop.
  * The things are grouped into categories, as specified by the categories props.
  */
-const SummaryList: React.FC<Props> = ({
-  heading,
-  items,
-  categories,
-  onChange,
-  onBlur,
-  colorSchema,
-  answers,
-  validationErrors,
-  showSum,
-  startEditable,
-  help,
-  editable,
-}) => {
+const SummaryList = (props: Props): JSX.Element | null => {
+  const {
+    heading,
+    items = [],
+    categories = [],
+    onBlur,
+    colorSchema,
+    answers,
+    validationErrors,
+    showSum = true,
+    startEditable,
+    help,
+    editable,
+    onChange = () => undefined,
+  } = props;
+
   const validColorSchema = getValidColorSchema(colorSchema);
 
   const sortedAnswers = useMemo(() => {
@@ -249,7 +144,7 @@ const SummaryList: React.FC<Props> = ({
   let sum = 0;
   const addToSum = (value: string | number) => {
     if (typeof value === "string") {
-      const summand = parseInt(value);
+      const summand = parseInt(value, 10);
       // eslint-disable-next-line no-restricted-globals
       sum += isNaN(summand) ? 0 : summand;
     } else {
@@ -445,7 +340,6 @@ const SummaryList: React.FC<Props> = ({
       listItems.push(
         <SummarySpacer
           key={`${listEntry.item.id}`}
-          category={listEntry.item.category}
           colorSchema={validColorSchema}
         />
       );
@@ -457,94 +351,34 @@ const SummaryList: React.FC<Props> = ({
   ).length;
 
   if (spacerCount > 1) {
-    const lastSpacerIndex = findLastIndex(listItems, (item) =>
-      item?.key?.toString().startsWith("spacer-")
+    const lastSpacerIndex = findLastIndex(
+      listItems,
+      (item) => !!item?.key?.toString().startsWith("spacer-")
     );
 
     listItems.splice(lastSpacerIndex, 1);
   }
 
-  return (
-    listItems.length > 0 && (
-      <>
-        <GroupedList
-          heading={heading}
-          categories={categories}
-          colorSchema={validColorSchema}
-          showEditButton={editable}
-          startEditable={startEditable && editable}
-          help={help}
-        >
-          {listItems}
-        </GroupedList>
-        {showSum && (
-          <SumContainer colorSchema={validColorSchema}>
-            <SumLabel colorSchema={validColorSchema}>Summa</SumLabel>
-            <SumText type="h1">{sum} kr</SumText>
-          </SumContainer>
-        )}
-      </>
-    )
-  );
-};
-
-SummaryList.propTypes = {
-  /**
-   * The header text of the list.
-   */
-  heading: PropTypes.string,
-  /**
-   * List of all items, corresponding to all subforms
-   */
-  items: PropTypes.array,
-  /**
-   * The categories of the grouping
-   */
-  categories: PropTypes.array,
-  /**
-   * What should happen to update the values
-   */
-  onChange: PropTypes.func,
-  /**
-   * What should happen when a field loses focus.
-   */
-  onBlur: PropTypes.func,
-  /**
-   * Sets the color scheme of the list. default is red.
-   */
-  color: PropTypes.string,
-  /**
-   * The form state answers
-   */
-  answers: PropTypes.object,
-  /**
-   * Object containing all validation errors for the entire form
-   */
-  validationErrors: PropTypes.object,
-  /**
-   * Whether or not to show a sum of all numeric values at the bottom. Defaults to true.
-   */
-  showSum: PropTypes.bool,
-  /**
-   * Whether to start in editable mode or not.
-   */
-  startEditable: PropTypes.bool,
-  /**
-   * Show a help button
-   */
-  help: PropTypes.shape({
-    text: PropTypes.string,
-    size: PropTypes.number,
-    heading: PropTypes.string,
-    tagline: PropTypes.string,
-    url: PropTypes.string,
-  }),
-};
-SummaryList.defaultProps = {
-  items: [],
-  color: "blue",
-  showSum: true,
-  onChange: () => {},
+  return listItems.length > 0 ? (
+    <>
+      <GroupedList
+        heading={heading}
+        categories={categories}
+        colorSchema={validColorSchema}
+        showEditButton={editable}
+        startEditable={startEditable && editable}
+        help={help}
+      >
+        {listItems}
+      </GroupedList>
+      {showSum && (
+        <SumContainer colorSchema={validColorSchema}>
+          <SumLabel colorSchema={validColorSchema}>Summa</SumLabel>
+          <SumText type="h1">{sum} kr</SumText>
+        </SumContainer>
+      )}
+    </>
+  ) : null;
 };
 
 export default SummaryList;
